+++ CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [Unreleased]
+### Miscellaneous
+
+- add automated changelog with git-cliff and pre-commit hook
 
 ## [0.1.0] - 2026-02-15